requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ✅ Supabase & dependencies (latest, proxy bug fix)
supabase>=2.4.0
httpx>=0.28.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.0.0

# ✅ Redis for multi-tier caching (enhanced versions)
redis==5.0.8

# OpenSearch client
opensearch-py==2.0.0
opensearch-py[async]==2.0.0
aiohttp==3.9.0   # ये dependency जरूरी है async ke liye

# ✅ Performance enhancements
orjson>=3.9.0  # Ultra-fast JSON serialization
email-validator
jinja2